==> Codes/00.Run All.R <==

# Chapter III - Public survey pilot data analysis

library(pacman)
library(tidyverse)
library(apollo)
library(car)
library(haven)
library(readxl)
library(tidyr)
library(stringr)
library(readr)
library(purrr)
library(cansim)
library(janitor)
library(anesrake)
library(naniar)
library(patchwork)
library(sf)
library(tmap)
library(tidygeocoder)

# create supportig data files to set the choice data set for Apollo - part 1
source("./Codes/Survey Data Preparation Part 1.R")

# create supportig data files to set the choice data set for Apollo - part 2
source("./Codes/Survey Data Preparation Part 2.R")

# create supportig data files to set the choice data set for Apollo - part 3
source("./codes/Survey Data Preparation Part 3.R")

# create supportig data files to set the choice data set for Apollo - part 4
source("./Codes/Survey Data Preparation Part 4.R")

# create supportig data files to set the choice data set for Apollo - part 5
source("./Codes/Survey Data Preparation Part 5.R")

# Manuscript Figure 1
source("./Codes/Figure 1.R")

# Manuscript Figure 3
source("./Codes/Figure 3.R")

# Manuscript Model 1 Estimation
source("./Codes/Model 1.R")

# Manuscript Model 2 Estimation
source("./Codes/Model 2.R")

# Manuscript Figure 5
source("./Codes/Figure 5.R")

# Manuscript Figure 6
source("./Codes/Figure 6.R")

# Manuscript Model 3 Estimation
source("./Codes/Model 3.R")

# Manuscript Model 4 Estimation
source("./Codes/Model 4.R")

# Manuscript Figure 7
source("./Codes/Figure 7.R")

# Manuscript Model 5 Estimation
source("./Codes/Model 5.R")

# Manuscript Model 6 Estimation
source("./Codes/Model 6.R")

# Manuscript Model 7 Estimation
source("./Codes/Model 7.R")

# Manuscript Figure 8
source("./Codes/Figure 8.R")


##############################################################################
